Let's wait with forcing clearlooks on everything.
Let's start styling as much as we can without resorting to theme engines and only when a widget is too complex to get away with borders and gradient fills and border-radius, we load the engine for specific widgets.

Probably started with teh wrong widget as GtkScrollbar seems to expose its properties in a really weird way.