adwaita: first pass of a big rewrite for borders and linked areas
We want to reduce the number of SVG assets we use for borders, and at
the same time add some outset shadows to elements using border-image.

Two new CSS files are created and included:
- gtk-widgets-backgrounds.css: defines the background gradients for
  common widgets such as entries, buttons and comboboxes
- gtk-widgets-borders.css: defines the borders for such common elements

This is a first pass of implementing this new style for the light theme
(dark theme support is still missing).