progress-bar: fix trough border rendering for selected cell renderers
For selected cell renderers with a trough component, we shouldn't be
using a border with a bottom highlight. We have the same problem with
check/radios, where we use an alternate asset. Do the same here.
Also, make the progressbar border a bit darker in that situation, so it
stands out from the selection color background.

Reported in https://bugzilla.gnome.org/show_bug.cgi?id=673321